+++ crates/larql-kv/src/engines/apollo/mod.rs
@@ -1,5 +1,6 @@
 pub mod engine;
 pub mod entry;
+pub(crate) mod executor;
 pub mod npy;
 pub mod routing;
 pub mod store;
