+++ crates/larql-kv/src/engines/boundary_per_layer/engine.rs
@@ -1,28 +1,33 @@
-//! `BoundaryPerLayerEngine` — `KvEngine` implementation with per-layer codec
-//! policy.
+//! `BoundaryPerLayerEngine` — `KvEngine` implementation with per-layer
+//! codec policy on the cold tier.
 //!
 //! The engine refuses to construct without a matching calibration record
-//! (per spec §4.7 + §4.9). v0.1 supports `Bf16` per layer only; other codec
-//! choices are rejected at policy construction (per
+//! (per spec §4.7 + §4.9). v0.1 supports `Bf16` per layer only; other
+//! codec choices are rejected at policy construction (per
 //! [`super::policy::PolicyError`]).
+//!
+//! Implementation is split across sibling modules:
+//!
+//! - this file: struct + construction + `KvEngine` trait glue
+//! - [`super::walk`] — CPU dense walk path (`run_prefill`/`run_decode`)
+//! - [`super::dispatch`] — W1-GPU dispatch fast path
+//!   (`try_prefill_via_dispatch`/`decode_step_via_dispatch`)
+//! - [`super::executor`] — `LayerExecutor`-driven path
+//!   (`prefill_via_executor`/`decode_step_via_executor`)
+//! - [`super::cold_tier`] — cold-tier maintenance
+//!   (`extend_cold_kv_with_overflow` + small helpers)
 
-use larql_compute::ComputeBackend;
-use larql_inference::attention::{
-    run_attention_block_decode_step_backend, run_attention_with_kv_backend, SharedKV,
-};
-use larql_inference::ffn::{BackendFfn, FfnBackend};
-use larql_inference::forward::{embed_tokens_pub, run_ffn};
+use larql_inference::ffn::FfnBackend;
 use larql_inference::model::ModelWeights;
 use larql_inference::{cpu_engine_backend, EngineBackend};
-use ndarray::{s, Array2};
+use ndarray::Array2;
 
 use crate::engines::boundary_per_layer::calibration::{
     BoundaryCalibrationRecord, BoundaryCalibrationStore, CalibrationError,
 };
 use crate::engines::boundary_per_layer::policy::BoundaryLayerPolicy;
-use crate::engines::boundary_per_layer::store::{PerLayerEncodedColdLayer, RsStorePerLayer};
-use crate::engines::markov_residual::recompute_kv;
-use crate::engines::markov_residual_codec::codec::ColdResidualCodec;
+use crate::engines::boundary_per_layer::store::RsStorePerLayer;
+use crate::engines::boundary_per_layer::{dispatch, executor, walk};
 use crate::{EngineInfo, KvEngine};
 
 /// Errors during engine construction (preconditions per spec §4.6).
@@ -39,11 +44,16 @@ pub enum EngineConstructionError {
 
 /// `BoundaryPerLayerEngine` — per-layer codec policy on the cold tier.
 pub struct BoundaryPerLayerEngine {
-    window_size: Option<usize>,
-    policy: BoundaryLayerPolicy,
-    record: BoundaryCalibrationRecord,
-    store: Option<RsStorePerLayer>,
-    backend: Box<dyn EngineBackend>,
+    pub(super) window_size: Option<usize>,
+    pub(super) policy: BoundaryLayerPolicy,
+    pub(super) record: BoundaryCalibrationRecord,
+    pub(super) store: Option<RsStorePerLayer>,
+    /// W1-GPU dispatch handle. `Some` when the prefill went through
+    /// `dispatch::try_prefill_via_dispatch` and decode is using the
+    /// kernel-fused fast path. `None` when the engine fell back to the
+    /// dense walk (e.g. backend lacks cached_decode support).
+    pub(super) kv_handle: Option<larql_inference::KvHandle>,
+    pub(super) backend: Box<dyn EngineBackend>,
 }
 
 impl BoundaryPerLayerEngine {
@@ -87,6 +97,7 @@ impl BoundaryPerLayerEngine {
             policy,
             record,
             store: None,
+            kv_handle: None,
             backend,
         })
     }
@@ -97,250 +108,6 @@ impl BoundaryPerLayerEngine {
 
     pub fn calibration_record(&self) -> &BoundaryCalibrationRecord {
         &self.record
-    }
-
-    fn run_prefill(&mut self, weights: &ModelWeights, token_ids: &[u32]) -> Option<Array2<f32>> {
-        let backend = self.backend.as_ref();
-        let num_layers = weights.num_layers;
-        let seq_len = token_ids.len();
-        let mut h = embed_tokens_pub(weights, token_ids);
-        let mut stored: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-        let be = Some(backend as &dyn ComputeBackend);
-
-        for layer in 0..num_layers {
-            stored.push(h.clone());
-            let (h_post_attn, _k, _v) =
-                run_attention_with_kv_backend(weights, &h, layer, be).expect("attention failed");
-            let bffn = BackendFfn {
-                weights,
-                backend: backend as &dyn ComputeBackend,
-            };
-            let (h_out, _) = run_ffn(weights, &h_post_attn, layer, &bffn, false);
-            h = h_out;
-        }
-
-        let mut rs = RsStorePerLayer {
-            stored,
-            cold_encoded: None,
-            cold_kv: None,
-            cold_abs_start: 0,
-            next_position: seq_len,
-            max_window: self.window_size,
-            policy_codecs: self.policy.entries.clone(),
-        };
-
-        let mut overflow_per_layer: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-        for layer in 0..num_layers {
-            overflow_per_layer.push(rs.clip_layer_overflow(layer));
-        }
-        if overflow_per_layer.first().map_or(0, |c| c.shape()[0]) > 0 {
-            let mut encoded_layers: Vec<PerLayerEncodedColdLayer> = Vec::with_capacity(num_layers);
-            let mut cold_kv: Vec<SharedKV> = Vec::with_capacity(num_layers);
-            for (layer, overflow) in overflow_per_layer.iter().enumerate() {
-                let codec = self.policy.codec_for(layer);
-                let decoded_overflow = roundtrip(overflow, codec);
-                let (k, v) = recompute_kv(weights, &decoded_overflow, layer, 0, backend, None)
-                    .expect("cold K/V pre-computation failed");
-                cold_kv.push((k, v));
-                let mut enc = PerLayerEncodedColdLayer::empty(codec, weights.hidden_size);
-                enc.append(overflow);
-                encoded_layers.push(enc);
-            }
-            rs.cold_encoded = Some(encoded_layers);
-            rs.cold_kv = Some(cold_kv);
-            rs.cold_abs_start = 0;
-        }
-
-        let last = last_row(&h);
-        self.store = Some(rs);
-        Some(last)
-    }
-
-    fn run_decode(&mut self, weights: &ModelWeights, token_id: u32) -> Option<Array2<f32>> {
-        let backend = self.backend.as_ref();
-        let mut rs = self.store.take()?;
-        let num_layers = weights.num_layers;
-        let abs_position = rs.next_position;
-        let mut h_new = embed_tokens_pub(weights, &[token_id]);
-        let mut new_stored: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-
-        for layer in 0..num_layers {
-            let h_hot = &rs.stored[layer];
-            let s_hot = h_hot.shape()[0];
-            let hot_abs_start = abs_position.saturating_sub(s_hot);
-
-            let (k_full, v_full) = if let Some(cold_kv) = &rs.cold_kv {
-                let (k_cold, v_cold) = &cold_kv[layer];
-                let (k_hot, v_hot) =
-                    recompute_kv(weights, h_hot, layer, hot_abs_start, backend, None)?;
-                let c = k_cold.shape()[0];
-                let kv_dim = k_cold.shape()[1];
-                let mut k_combined = Array2::<f32>::zeros((c + s_hot, kv_dim));
-                k_combined.slice_mut(s![..c, ..]).assign(k_cold);
-                k_combined.slice_mut(s![c.., ..]).assign(&k_hot);
-                let mut v_combined = Array2::<f32>::zeros((c + s_hot, kv_dim));
-                v_combined.slice_mut(s![..c, ..]).assign(v_cold);
-                v_combined.slice_mut(s![c.., ..]).assign(&v_hot);
-                (k_combined, v_combined)
-            } else {
-                let (h_full, full_abs_start) = if let Some(cold_layers) = &rs.cold_encoded {
-                    let enc = &cold_layers[layer];
-                    if enc.n_positions > 0 {
-                        let decoded = enc.decode();
-                        let hidden = h_hot.shape()[1];
-                        let mut combined =
-                            Array2::<f32>::zeros((decoded.shape()[0] + s_hot, hidden));
-                        combined
-                            .slice_mut(s![..decoded.shape()[0], ..])
-                            .assign(&decoded);
-                        combined
-                            .slice_mut(s![decoded.shape()[0].., ..])
-                            .assign(h_hot);
-                        (combined, rs.cold_abs_start)
-                    } else {
-                        (h_hot.clone(), hot_abs_start)
-                    }
-                } else {
-                    (h_hot.clone(), hot_abs_start)
-                };
-                let (k, v) = recompute_kv(weights, &h_full, layer, full_abs_start, backend, None)?;
-                (k, v)
-            };
-
-            new_stored.push(h_new.clone());
-
-            let (h_post_attn, _new_kv) = run_attention_block_decode_step_backend(
-                weights,
-                &h_new,
-                layer,
-                Some(&(k_full, v_full)),
-                abs_position,
-                Some(backend as &dyn ComputeBackend),
-            )?;
-
-            let bffn = BackendFfn {
-                weights,
-                backend: backend as &dyn ComputeBackend,
-            };
-            let (h_out, _) = run_ffn(weights, &h_post_attn, layer, &bffn, false);
-            h_new = h_out;
-        }
-
-        // Amortised O(m) per-row append via ndarray::Array2::push_row.
-        // Replaces the O(N²) per-step "Array2::zeros + .assign" rebuild.
-        for (slab, new_row) in rs.stored.iter_mut().zip(new_stored.iter()) {
-            slab.push_row(new_row.row(0))
-                .expect("push_row shape mismatch");
-        }
-
-        let mut updated_rs = RsStorePerLayer {
-            stored: rs.stored,
-            cold_encoded: rs.cold_encoded,
-            cold_kv: rs.cold_kv,
-            cold_abs_start: rs.cold_abs_start,
-            next_position: abs_position + 1,
-            max_window: rs.max_window,
-            policy_codecs: rs.policy_codecs,
-        };
-
-        let mut overflow_per_layer: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-        for layer in 0..num_layers {
-            overflow_per_layer.push(updated_rs.clip_layer_overflow(layer));
-        }
-        if overflow_per_layer.first().map_or(0, |c| c.shape()[0]) > 0 {
-            // Snapshot the absolute position at which the new overflow rows land
-            // BEFORE appending to cold_encoded. Used by extend_cold_kv for RoPE.
-            let cold_abs_pos = updated_rs.cold_abs_start
-                + updated_rs
-                    .cold_encoded
-                    .as_ref()
-                    .map_or(0, |l| l[0].n_positions);
-            match updated_rs.cold_encoded.as_mut() {
-                Some(layers) => {
-                    for (layer, overflow) in overflow_per_layer.iter().enumerate() {
-                        layers[layer].append(overflow);
-                    }
-                }
-                None => {
-                    let hidden = weights.hidden_size;
-                    let mut layers: Vec<PerLayerEncodedColdLayer> = Vec::with_capacity(num_layers);
-                    for (layer, overflow) in overflow_per_layer.iter().enumerate() {
-                        let codec = self.policy.codec_for(layer);
-                        let mut enc = PerLayerEncodedColdLayer::empty(codec, hidden);
-                        enc.append(overflow);
-                        layers.push(enc);
-                    }
-                    updated_rs.cold_encoded = Some(layers);
-                }
-            }
-            self.extend_cold_kv_with_overflow(
-                weights,
-                backend,
-                &mut updated_rs,
-                &overflow_per_layer,
-                cold_abs_pos,
-            );
-        }
-
-        let last = last_row(&h_new);
-        self.store = Some(updated_rs);
-        Some(last)
-    }
-
-    /// Extend `cold_kv` to cover newly-evicted overflow rows.
-    ///
-    /// Computes K/V on the codec round-trip of each layer's overflow
-    /// (preserving the lossy contract used at prefill) and concatenates
-    /// onto each layer's existing cold (K, V). If `cold_kv` is `None`,
-    /// initialises it. Replaces the previous "nuke cold_kv on every
-    /// overflow" path which forced the next decode step to recompute
-    /// K/V over the entire cold tier — O(N²) in windowed mode.
-    fn extend_cold_kv_with_overflow(
-        &self,
-        weights: &ModelWeights,
-        backend: &dyn ComputeBackend,
-        rs: &mut RsStorePerLayer,
-        overflow_per_layer: &[Array2<f32>],
-        cold_abs_pos: usize,
-    ) -> Option<()> {
-        let num_layers = weights.num_layers;
-        let n_new = overflow_per_layer.first().map_or(0, |c| c.shape()[0]);
-        if n_new == 0 {
-            return Some(());
-        }
-        match rs.cold_kv.as_mut() {
-            Some(cold_kv) => {
-                for (layer, overflow) in overflow_per_layer.iter().enumerate() {
-                    let codec = self.policy.codec_for(layer);
-                    let decoded = roundtrip(overflow, codec);
-                    let (k_new, v_new) =
-                        recompute_kv(weights, &decoded, layer, cold_abs_pos, backend, None)?;
-                    let (k_old, v_old) = &cold_kv[layer];
-                    let kv_dim = k_old.shape()[1];
-                    let l_old = k_old.shape()[0];
-                    let l_total = l_old + n_new;
-                    let mut k = Array2::<f32>::zeros((l_total, kv_dim));
-                    k.slice_mut(s![..l_old, ..]).assign(k_old);
-                    k.slice_mut(s![l_old.., ..]).assign(&k_new);
-                    let mut v = Array2::<f32>::zeros((l_total, kv_dim));
-                    v.slice_mut(s![..l_old, ..]).assign(v_old);
-                    v.slice_mut(s![l_old.., ..]).assign(&v_new);
-                    cold_kv[layer] = (k, v);
-                }
-            }
-            None => {
-                let mut new_cold_kv: Vec<SharedKV> = Vec::with_capacity(num_layers);
-                for (layer, overflow) in overflow_per_layer.iter().enumerate() {
-                    let codec = self.policy.codec_for(layer);
-                    let decoded = roundtrip(overflow, codec);
-                    let (k, v) =
-                        recompute_kv(weights, &decoded, layer, cold_abs_pos, backend, None)?;
-                    new_cold_kv.push((k, v));
-                }
-                rs.cold_kv = Some(new_cold_kv);
-            }
-        }
-        Some(())
     }
 }
 
@@ -370,19 +137,38 @@ impl KvEngine for BoundaryPerLayerEngine {
     fn prefill(
         &mut self,
         weights: &ModelWeights,
-        _ffn: &dyn FfnBackend,
+        ffn: &dyn FfnBackend,
         token_ids: &[u32],
     ) -> Option<Array2<f32>> {
-        self.run_prefill(weights, token_ids)
+        let (hidden, store) = walk::run_prefill(
+            weights,
+            ffn,
+            self.backend.as_ref(),
+            &self.policy,
+            self.window_size,
+            token_ids,
+        )?;
+        self.store = Some(store);
+        Some(hidden)
     }
 
     fn decode_step(
         &mut self,
         weights: &ModelWeights,
-        _ffn: &dyn FfnBackend,
+        ffn: &dyn FfnBackend,
         token_id: u32,
     ) -> Option<Array2<f32>> {
-        self.run_decode(weights, token_id)
+        let rs = self.store.take()?;
+        let (hidden, new_rs) = walk::run_decode(
+            weights,
+            ffn,
+            self.backend.as_ref(),
+            &self.policy,
+            rs,
+            token_id,
+        )?;
+        self.store = Some(new_rs);
+        Some(hidden)
     }
 
     fn memory_bytes(&self) -> usize {
@@ -399,12 +185,9 @@ impl KvEngine for BoundaryPerLayerEngine {
 
     // ── Q4K path ─────────────────────────────────────────────────────────
     //
-    // boundary_per_layer doesn't yet have a fused-quant kernel route;
-    // it dequantises Q4K attention tensors into `weights.tensors` via
-    // the standard helper, then delegates to the f32 dense walk. This
-    // unblocks benching on Q4K-only vindexes (and matches what
-    // markov-rs-codec did before its W1-GPU `try_prefill_via_dispatch`
-    // shortcut landed — boundary_per_layer can pick that up later).
+    // Try W1-GPU dispatch first; fall back to dense walk with attn
+    // tensors dequantised when the backend / vindex doesn't support
+    // direct-matvec decode.
 
     fn prefill_quant(
         &mut self,
@@ -414,6 +197,20 @@ impl KvEngine for BoundaryPerLayerEngine {
         token_ids: &[u32],
         _backend: &dyn larql_compute::ComputeBackend,
     ) -> Option<Array2<f32>> {
+        if let Some((hidden, store, handle)) = dispatch::try_prefill_via_dispatch(
+            weights,
+            self.backend.as_ref(),
+            &self.policy,
+            self.window_size,
+            index,
+            token_ids,
+        ) {
+            self.store = Some(store);
+            self.kv_handle = Some(handle);
+            return Some(hidden);
+        }
+        // Fall back to dense f32 walk (compact vindexes / CPU backend).
+        self.kv_handle = None;
         larql_inference::vindex::dequant::ensure_attn_tensors_dequantised(weights, index);
         self.prefill(weights, ffn, token_ids)
     }
@@ -426,15 +223,42 @@ impl KvEngine for BoundaryPerLayerEngine {
         token_id: u32,
         _backend: &dyn larql_compute::ComputeBackend,
     ) -> Option<Array2<f32>> {
+        // If prefill went through dispatch, decode does too.
+        if self.kv_handle.is_some() {
+            let mut handle = self.kv_handle.take()?;
+            let rs = self.store.take()?;
+            let result = dispatch::decode_step_via_dispatch(
+                weights,
+                self.backend.as_ref(),
+                &self.policy,
+                &mut handle,
+                rs,
+                index,
+                token_id,
+            );
+            match result {
+                Some((hidden, new_rs)) => {
+                    self.store = Some(new_rs);
+                    self.kv_handle = Some(handle);
+                    return Some(hidden);
+                }
+                None => {
+                    // State-dump failure — clear handle, fall through to
+                    // dense walk on the next call.
+                    self.kv_handle = None;
+                    return None;
+                }
+            }
+        }
         larql_inference::vindex::dequant::ensure_attn_tensors_dequantised(weights, index);
         self.decode_step(weights, ffn, token_id)
     }
 
     // ── Phase 2 migration: executor-driven path ──────────────────────────
     //
-    // Per-layer codec policy requires per-layer dispatch. Override the
-    // dense (non-quant) via_executor methods to drive the layer loop
-    // through the executor + honor the caller's FFN backend.
+    // Per-layer codec policy requires per-layer dispatch. The executor
+    // path drives the layer loop through a caller-supplied executor +
+    // honours the caller's FFN backend.
 
     fn prefill_via_executor(
         &mut self,
@@ -443,61 +267,21 @@ impl KvEngine for BoundaryPerLayerEngine {
         ffn: &dyn FfnBackend,
         token_ids: &[u32],
     ) -> Option<Array2<f32>> {
-        use crate::engines::markov_residual::recompute_kv;
         use larql_inference::layer_executor::ExecutorDispatchKind;
-
         if matches!(executor.dispatch_kind(), ExecutorDispatchKind::Fused) {
             // State policy can't fire under fused dispatch; degrade.
             return self.prefill(weights, ffn, token_ids);
         }
-
-        let backend = executor.backend();
-        let num_layers = weights.num_layers;
-        let seq_len = token_ids.len();
-        let mut h = embed_tokens_pub(weights, token_ids);
-        let mut stored: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-
-        for layer in 0..num_layers {
-            stored.push(h.clone());
-            let (h_out, _kv) = executor.run_prefill_layer(weights, layer, &h, ffn)?;
-            h = h_out;
-        }
-
-        let mut rs = RsStorePerLayer {
-            stored,
-            cold_encoded: None,
-            cold_kv: None,
-            cold_abs_start: 0,
-            next_position: seq_len,
-            max_window: self.window_size,
-            policy_codecs: self.policy.entries.clone(),
-        };
-
-        let mut overflow_per_layer: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-        for layer in 0..num_layers {
-            overflow_per_layer.push(rs.clip_layer_overflow(layer));
-        }
-        if overflow_per_layer.first().map_or(0, |c| c.shape()[0]) > 0 {
-            let mut encoded_layers: Vec<PerLayerEncodedColdLayer> = Vec::with_capacity(num_layers);
-            let mut cold_kv: Vec<SharedKV> = Vec::with_capacity(num_layers);
-            for (layer, overflow) in overflow_per_layer.iter().enumerate() {
-                let codec = self.policy.codec_for(layer);
-                let decoded_overflow = roundtrip(overflow, codec);
-                let (k, v) = recompute_kv(weights, &decoded_overflow, layer, 0, backend, None)
-                    .expect("cold K/V pre-computation failed");
-                cold_kv.push((k, v));
-                let mut enc = PerLayerEncodedColdLayer::empty(codec, weights.hidden_size);
-                enc.append(overflow);
-                encoded_layers.push(enc);
-            }
-            rs.cold_encoded = Some(encoded_layers);
-            rs.cold_kv = Some(cold_kv);
-            rs.cold_abs_start = 0;
-        }
-
-        let out = last_row(&h);
-        self.store = Some(rs);
-        Some(out)
+        let (hidden, store) = executor::run_prefill(
+            weights,
+            executor,
+            ffn,
+            &self.policy,
+            self.window_size,
+            token_ids,
+        )?;
+        self.store = Some(store);
+        Some(hidden)
     }
 
     fn decode_step_via_executor(
@@ -507,135 +291,16 @@ impl KvEngine for BoundaryPerLayerEngine {
         ffn: &dyn FfnBackend,
         token_id: u32,
     ) -> Option<Array2<f32>> {
-        use crate::engines::markov_residual::recompute_kv;
         use larql_inference::layer_executor::ExecutorDispatchKind;
-
         if matches!(executor.dispatch_kind(), ExecutorDispatchKind::Fused) {
             return self.decode_step(weights, ffn, token_id);
         }
-
-        let backend = executor.backend();
-        let mut rs = self.store.take()?;
-        let num_layers = weights.num_layers;
-        let abs_position = rs.next_position;
-        let mut h_new = embed_tokens_pub(weights, &[token_id]);
-        let mut new_stored: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-
-        for layer in 0..num_layers {
-            let h_hot = &rs.stored[layer];
-            let s_hot = h_hot.shape()[0];
-            let hot_abs_start = abs_position.saturating_sub(s_hot);
-
-            let prior_kv: SharedKV = if let Some(cold_kv) = &rs.cold_kv {
-                let (k_cold, v_cold) = &cold_kv[layer];
-                let (k_hot, v_hot) =
-                    recompute_kv(weights, h_hot, layer, hot_abs_start, backend, None)?;
-                let c = k_cold.shape()[0];
-                let kv_dim = k_cold.shape()[1];
-                let mut k_combined = Array2::<f32>::zeros((c + s_hot, kv_dim));
-                k_combined.slice_mut(s![..c, ..]).assign(k_cold);
-                k_combined.slice_mut(s![c.., ..]).assign(&k_hot);
-                let mut v_combined = Array2::<f32>::zeros((c + s_hot, kv_dim));
-                v_combined.slice_mut(s![..c, ..]).assign(v_cold);
-                v_combined.slice_mut(s![c.., ..]).assign(&v_hot);
-                (k_combined, v_combined)
-            } else {
-                let (h_full, full_abs_start) = match &rs.cold_encoded {
-                    Some(cold_layers) if cold_layers[layer].n_positions > 0 => {
-                        let decoded = cold_layers[layer].decode();
-                        let hidden = h_hot.shape()[1];
-                        let mut combined =
-                            Array2::<f32>::zeros((decoded.shape()[0] + s_hot, hidden));
-                        combined
-                            .slice_mut(s![..decoded.shape()[0], ..])
-                            .assign(&decoded);
-                        combined
-                            .slice_mut(s![decoded.shape()[0].., ..])
-                            .assign(h_hot);
-                        (combined, rs.cold_abs_start)
-                    }
-                    _ => (h_hot.clone(), hot_abs_start),
-                };
-                recompute_kv(weights, &h_full, layer, full_abs_start, backend, None)?
-            };
-
-            new_stored.push(h_new.clone());
-            let (h_out, _new_kv) =
-                executor.run_decode_layer(weights, layer, &h_new, &prior_kv, abs_position, ffn)?;
-            h_new = h_out;
-        }
-
-        // Amortised O(m) per-row append (see legacy decode for rationale).
-        for (slab, new_row) in rs.stored.iter_mut().zip(new_stored.iter()) {
-            slab.push_row(new_row.row(0))
-                .expect("push_row shape mismatch");
-        }
-
-        let mut updated_rs = RsStorePerLayer {
-            stored: rs.stored,
-            cold_encoded: rs.cold_encoded,
-            cold_kv: rs.cold_kv,
-            cold_abs_start: rs.cold_abs_start,
-            next_position: abs_position + 1,
-            max_window: rs.max_window,
-            policy_codecs: rs.policy_codecs,
-        };
-
-        let mut overflow_per_layer: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-        for layer in 0..num_layers {
-            overflow_per_layer.push(updated_rs.clip_layer_overflow(layer));
-        }
-        if overflow_per_layer.first().map_or(0, |c| c.shape()[0]) > 0 {
-            let cold_abs_pos = updated_rs.cold_abs_start
-                + updated_rs
-                    .cold_encoded
-                    .as_ref()
-                    .map_or(0, |l| l[0].n_positions);
-            match updated_rs.cold_encoded.as_mut() {
-                Some(layers) => {
-                    for (layer, overflow) in overflow_per_layer.iter().enumerate() {
-                        layers[layer].append(overflow);
-                    }
-                }
-                None => {
-                    let hidden = weights.hidden_size;
-                    let mut layers: Vec<PerLayerEncodedColdLayer> = Vec::with_capacity(num_layers);
-                    for (layer, overflow) in overflow_per_layer.iter().enumerate() {
-                        let codec = self.policy.codec_for(layer);
-                        let mut enc = PerLayerEncodedColdLayer::empty(codec, hidden);
-                        enc.append(overflow);
-                        layers.push(enc);
-                    }
-                    updated_rs.cold_encoded = Some(layers);
-                }
-            }
-            self.extend_cold_kv_with_overflow(
-                weights,
-                backend,
-                &mut updated_rs,
-                &overflow_per_layer,
-                cold_abs_pos,
-            );
-        }
-
-        let out = last_row(&h_new);
-        self.store = Some(updated_rs);
-        Some(out)
+        let rs = self.store.take()?;
+        let (hidden, new_rs) =
+            executor::run_decode(weights, executor, ffn, &self.policy, rs, token_id)?;
+        self.store = Some(new_rs);
+        Some(hidden)
     }
-}
-
-fn roundtrip(block: &Array2<f32>, codec: ColdResidualCodec) -> Array2<f32> {
-    if block.shape()[0] == 0 {
-        return block.clone();
-    }
-    let mut tmp = PerLayerEncodedColdLayer::empty(codec, block.shape()[1]);
-    tmp.append(block);
-    tmp.decode()
-}
-
-fn last_row(h: &Array2<f32>) -> Array2<f32> {
-    let last = h.shape()[0] - 1;
-    h.slice(s![last..=last, ..]).to_owned()
 }
 
 #[cfg(test)]
@@ -844,16 +509,12 @@ mod tests {
             BoundaryPerLayerEngine::new(Some(2), policy, weights.num_layers, &store).unwrap();
         eng.prefill(&weights, &ffn, &[0u32, 1, 2, 3])
             .expect("prefill");
-        // Prefill leaves cold_kv = Some (built from the prefill overflow).
         assert!(eng.store.as_ref().unwrap().cold_kv.is_some());
-        // First decode overflows → cold_kv must still be Some.
         eng.decode_step(&weights, &ffn, 4).expect("first decode");
         assert!(
             eng.store.as_ref().unwrap().cold_kv.is_some(),
             "cold_kv should stay Some after overflow (was being nuked pre-fix)"
         );
-        // Second decode overflows → cold_kv still Some, hits the cold_kv
-        // combine branch (not the cold_encoded recompute fallback).
         let h = eng.decode_step(&weights, &ffn, 5).expect("second decode");
         assert!(eng.store.as_ref().unwrap().cold_kv.is_some());
         assert_eq!(h.shape(), &[1, weights.hidden_size]);
@@ -876,26 +537,6 @@ mod tests {
         let m2 = eng.memory_bytes();
         assert!(m1 > m0);
         assert!(m2 > m1);
-    }
-
-    #[test]
-    fn roundtrip_empty_block_short_circuits() {
-        let empty: Array2<f32> = Array2::zeros((0, 8));
-        let out = roundtrip(&empty, ColdResidualCodec::Bf16);
-        assert_eq!(out.shape(), &[0, 8]);
-    }
-
-    #[test]
-    fn last_row_extracts_correct_row() {
-        let mut h = Array2::<f32>::zeros((3, 4));
-        for j in 0..4 {
-            h[[2, j]] = (j + 1) as f32;
-        }
-        let r = last_row(&h);
-        assert_eq!(r.shape(), &[1, 4]);
-        for j in 0..4 {
-            assert_eq!(r[[0, j]], (j + 1) as f32);
-        }
     }
 
     // ── Phase 2 migration: executor-driven path ──────────────────────────
@@ -1005,19 +646,12 @@ mod tests {
         engine
             .prefill(&weights, &ffn, &[0u32, 1, 2, 3])
             .expect("prefill overflow");
-        // First decode: cold_kv populated by prefill → hits combine branch.
-        // Overflow appends to cold_encoded + extend_cold_kv (Some branch).
         let h = engine.decode_step(&weights, &ffn, 4).expect("decode 1");
         assert_eq!(h.shape(), &[1, weights.hidden_size]);
-        // Second decode: cold_kv still Some (post-fix) → combine branch
-        // again; overflow extends cold_kv a second time.
         let h2 = engine.decode_step(&weights, &ffn, 5).expect("decode 2");
         assert_eq!(h2.shape(), &[1, weights.hidden_size]);
     }
 
-    /// Executor-driven decode with cold tier — exercises the same
-    /// cold_kv / cold_encoded branches in `decode_step_via_executor`
-    /// (lines 431-456 / 494-).
     #[test]
     fn decode_via_executor_traverses_cold_tier_branches() {
         use larql_inference::ffn::NullFfn;
@@ -1033,19 +667,17 @@ mod tests {
         engine
             .prefill_via_executor(&weights, &executor, &ffn, &[0u32, 1, 2, 3])
             .expect("prefill overflow");
-        // First decode: cold_kv combine branch.
         engine
             .decode_step_via_executor(&weights, &executor, &ffn, 4)
             .expect("decode 1");
-        // Second decode: cold_encoded recompute branch.
         let h = engine
             .decode_step_via_executor(&weights, &executor, &ffn, 5)
             .expect("decode 2");
         assert_eq!(h.shape(), &[1, weights.hidden_size]);
     }
 
-    /// Fused-executor fallback: lines 350-352 / 414-415 dispatch back
-    /// through the legacy `prefill` / `decode_step` path.
+    /// Fused-executor fallback: dispatches back through the legacy
+    /// `prefill` / `decode_step` path.
     struct FusedStubExecutor {
         backend: larql_compute::CpuBackend,
     }
