+++ crates/larql-kv/src/engines/markov_residual/engine.rs
@@ -2,57 +2,7 @@
 
 use larql_compute::ComputeBackend;
 use larql_vindex::VectorIndex;
-use ndarray::{s, Array2};
-
-// ── W8.2 helpers: pre-allocated doubling-capacity buffers ────────────────
-// Used by `try_prefill_via_dispatch` to seed `stored` / `hot_kv` and by
-// `decode_step_via_dispatch` to grow them. The hot path appends with one
-// `slice_mut(s![pos..pos+1, ..]).assign(row)` instead of allocating a
-// fresh `Array2::zeros((n+1, kv_dim))` each step — which the samply
-// flamegraph surfaced as 58% of decode CPU on the cached-state engines
-// (`__bzero` + `zip_mut_with_same_shape` + `madvise`).
-
-/// Initial doubling-capacity for `stored` / `hot_kv` given the prefill's
-/// `prompt_len` and the engine's optional sliding-window cap.
-fn window_capacity(prompt_len: usize, window_size: Option<usize>) -> usize {
-    match window_size {
-        Some(w) => prompt_len.max(w),
-        None => (prompt_len * 2).max(64),
-    }
-}
-
-/// Allocate an `[cap, cols]` Array2 and copy the first `len` rows from
-/// `src` (which is shape `[len, cols]`). Asserts `src.shape()[0] ==
-/// len`. Used at prefill to convert the captured `[prompt_len, dim]`
-/// state into the doubling-capacity layout.
-fn grow_capacity_2d(src: &Array2<f32>, len: usize, cap: usize) -> Array2<f32> {
-    debug_assert_eq!(src.shape()[0], len, "src shape disagrees with len");
-    debug_assert!(cap >= len, "cap {cap} smaller than len {len}");
-    let cols = src.shape()[1];
-    let mut buf = Array2::<f32>::zeros((cap, cols));
-    if len > 0 {
-        buf.slice_mut(s![..len, ..]).assign(src);
-    }
-    buf
-}
-
-/// Append one row to a pre-allocated doubling-capacity buffer. If the
-/// buffer is full (`len == cap`), doubles capacity, copies the live
-/// rows, and falls through to the in-place assign. `len` is the
-/// pre-append logical row count; caller increments it after.
-fn append_row(buf: &mut Array2<f32>, row: &Array2<f32>, len: usize) {
-    let cap = buf.shape()[0];
-    if len == cap {
-        let cols = buf.shape()[1];
-        let new_cap = (cap * 2).max(8);
-        let mut new_buf = Array2::<f32>::zeros((new_cap, cols));
-        new_buf
-            .slice_mut(s![..len, ..])
-            .assign(&buf.slice(s![..len, ..]));
-        *buf = new_buf;
-    }
-    buf.slice_mut(s![len..len + 1, ..]).assign(row);
-}
+use ndarray::Array2;
 
 use super::compute::{rs_decode_step, rs_decode_step_profiled, rs_prefill};
 use super::store::RsStore;
@@ -64,18 +14,18 @@ use larql_inference::model::ModelWeights;
 use larql_inference::{cpu_engine_backend, EngineBackend};
 
 pub struct MarkovResidualEngine {
-    window_size: Option<usize>,
-    store: Option<RsStore>,
-    backend: Box<dyn EngineBackend>,
-    profiling: bool,
-    profile: EngineProfiler,
+    pub(super) window_size: Option<usize>,
+    pub(super) store: Option<RsStore>,
+    pub(super) backend: Box<dyn EngineBackend>,
+    pub(super) profiling: bool,
+    pub(super) profile: EngineProfiler,
     /// W1-GPU: handle into the backend's internal K/V cache, populated
     /// when `prefill_quant` routes through `coarse_prefill_with_state`.
     /// `None` means the engine took the legacy per-layer walk path.
-    kv_handle: Option<larql_inference::KvHandle>,
+    pub(super) kv_handle: Option<larql_inference::KvHandle>,
     /// Position counter used by `coarse_decode_step_with_state` for RoPE.
     /// Tracks `prompt_len + steps_already_decoded`.
-    abs_position: usize,
+    pub(super) abs_position: usize,
 }
 
 impl MarkovResidualEngine {
@@ -103,336 +53,12 @@ impl MarkovResidualEngine {
     pub fn total_memory_bytes(&self) -> usize {
         self.store.as_ref().map_or(0, |s| s.memory_bytes())
     }
-
-    /// W1-GPU: attempt prefill through `KvDispatch::coarse_prefill_with_state`.
-    /// Returns `Some(hidden)` when the backend implements the GPU/fused path;
-    /// `None` when it doesn't (engine falls back to per-layer walk).
-    ///
-    /// On success: populates `self.store` from `state.h_in_per_layer` (the
-    /// residual store) and `state.k_new/v_new_per_layer` (the W2 hot K/V
-    /// cache, ready for the next decode step). Stashes the returned
-    /// `KvHandle` so `decode_step_via_dispatch` can continue on the same
-    /// fast path.
-    fn try_prefill_via_dispatch(
-        &mut self,
-        weights: &mut ModelWeights,
-        index: &VectorIndex,
-        token_ids: &[u32],
-    ) -> Option<Array2<f32>> {
-        use larql_inference::PerLayerDecodeState;
-        // Gate on Q4K vindex support — the dispatch path requires
-        // `insert_q4k_layer_tensors` to succeed for every layer, which
-        // means the vindex must carry Q4K attn data. Non-Q4K test
-        // fixtures (legacy synthetic vindex) fall through to the walk
-        // path. `supports_direct_matvec_decode` is the same gate
-        // `CpuBackend::coarse_decode_step` uses for routing.
-        if !larql_inference::vindex::supports_cached_decode(weights)
-            || !larql_inference::vindex::supports_direct_matvec_decode(weights, index)
-        {
-            return None;
-        }
-        let num_layers = weights.num_layers;
-        let mut state = PerLayerDecodeState::with_capacity(num_layers);
-        let (hidden, handle) = self.backend.as_ref().coarse_prefill_with_state(
-            weights,
-            token_ids,
-            Some(index),
-            Some(&mut state),
-        )?;
-        if !state.is_complete_for(num_layers) {
-            // Backend declined to dump per-layer state; can't drive the
-            // engine's contract from this path. Caller falls back.
-            return None;
-        }
-        // W8.2: pre-allocate `stored` and `hot_kv` to a doubling capacity
-        // so subsequent `decode_step_via_dispatch` calls can append
-        // in-place without re-allocating a fresh Array2 each step. The
-        // captured prefill state has shape `[prompt_len, dim]`; we copy
-        // it into a `[capacity, dim]` buffer where `capacity = max(2 *
-        // prompt_len, 64)`. The hot path doubles capacity on overflow.
-        let prompt_len = token_ids.len();
-        let initial_cap = window_capacity(prompt_len, self.window_size);
-        // W10 Phase A: consume each layer's handle into an owned
-        // Array2; CpuStateHandle moves without a copy.
-        let stored: Vec<Array2<f32>> = state
-            .h_in_per_layer
-            .into_iter()
-            .map(|h| grow_capacity_2d(&h.into_array(), prompt_len, initial_cap))
-            .collect();
-        let hot_kv: Vec<larql_inference::attention::SharedKV> = state
-            .k_new_per_layer
-            .into_iter()
-            .zip(state.v_new_per_layer)
-            .map(|(k, v)| {
-                (
-                    grow_capacity_2d(&k.into_array(), prompt_len, initial_cap),
-                    grow_capacity_2d(&v.into_array(), prompt_len, initial_cap),
-                )
-            })
-            .collect();
-        // W10 Phase B: when `LARQL_W10_HONLY=1`, drop the hot_kv
-        // shadow. Metal's own kv cache is the source of truth for K/V
-        // on this engine (markov_residual treats K/V as derivative —
-        // see crates/larql-kv/docs/state-policy.md §2.2). Decode steps
-        // can then request the HOnly capture mask, skipping the K/V
-        // staging buffer alloc + GPU→CPU readback.
-        //
-        // W10 Phase C: when window_size is also None (no cold-tier
-        // eviction can ever fire), `rs.stored` is dead weight too —
-        // nothing reads it after prefill. Drop it; decode steps can
-        // then request the None mask, eliminating the h_in staging
-        // and readback alongside K/V.
-        let drop_hot_kv_shadow = std::env::var("LARQL_W10_HONLY")
-            .ok()
-            .map(|v| v == "1")
-            .unwrap_or(false);
-        let drop_stored_shadow = drop_hot_kv_shadow && self.window_size.is_none();
-        let stored = if drop_stored_shadow {
-            // Empty per-layer placeholders — memory_bytes() honestly
-            // reports ~0 and recompute_kv is unreachable (no window
-            // → no cold-tier eviction → no replay).
-            let hidden_size = weights.hidden_size;
-            (0..num_layers)
-                .map(|_| Array2::<f32>::zeros((0, hidden_size)))
-                .collect()
-        } else {
-            stored
-        };
-        let mut rs = RsStore {
-            stored,
-            cold_residuals: None,
-            cold_kv: None,
-            cold_len: 0,
-            hot_kv: if drop_hot_kv_shadow {
-                None
-            } else {
-                Some(hot_kv)
-            },
-            cold_abs_start: 0,
-            next_position: prompt_len,
-            max_window: self.window_size,
-            hot_len: if drop_stored_shadow { 0 } else { prompt_len },
-        };
-        // Clip window on prefill — overflow goes into cold tier using
-        // the snapshot helper (already-computed K/V from the dispatch).
-        // W8.2: use `rs.hot_len`, not `stored[l].shape()[0]` (pre-alloc
-        // capacity).
-        let pre_clip: Vec<usize> = if rs.hot_kv.is_some() {
-            let window = self.window_size.unwrap_or(usize::MAX);
-            let evict_count = rs.hot_len.saturating_sub(window);
-            vec![evict_count; rs.stored.len()]
-        } else {
-            Vec::new()
-        };
-        let evicted_hot_kv = rs
-            .hot_kv
-            .as_ref()
-            .filter(|_| pre_clip.iter().any(|&n| n > 0))
-            .and_then(|h| RsStore::snapshot_evicted_hot_kv(h, &pre_clip));
-        let mut cold: Vec<ndarray::Array2<f32>> = Vec::with_capacity(num_layers);
-        for layer in 0..num_layers {
-            rs.clip_layer(layer, &mut cold);
-        }
-        rs.finalise_hot_len_after_clip();
-        // 2026-05-19 audit fix: route through the doubling-capacity
-        // append helper so cold_len is correctly initialised. Subsequent
-        // overflow-merge calls on the same store stay amortised O(1).
-        rs.append_cold_overflow(cold, evicted_hot_kv);
-        if rs.cold_len > 0 {
-            rs.cold_abs_start = 0;
-        }
-        self.store = Some(rs);
-        self.kv_handle = Some(handle);
-        self.abs_position = token_ids.len();
-        Some(hidden)
-    }
-
-    /// W1-GPU: decode step through `KvDispatch::coarse_decode_step_with_state`.
-    /// Caller guarantees `self.kv_handle` is `Some` (set by
-    /// `try_prefill_via_dispatch`); per-layer state is appended to the
-    /// engine's store/hot_kv on each step.
-    fn decode_step_via_dispatch(
-        &mut self,
-        weights: &mut ModelWeights,
-        index: &VectorIndex,
-        token_id: u32,
-    ) -> Option<Array2<f32>> {
-        // W10 instrumentation: decode_total wraps the whole step so
-        // stage_summary returns Some on the dispatch hot path (the
-        // legacy walk path records it separately at the bottom of
-        // rs_decode_step_profiled).
-        let t_total = std::time::Instant::now();
-        use larql_inference::PerLayerDecodeState;
-        use ndarray::Array2;
-        let num_layers = weights.num_layers;
-        let mut state = PerLayerDecodeState::with_capacity(num_layers);
-        let handle = self.kv_handle.as_mut()?;
-        // W10 Phase B/C: select capture mask. When `LARQL_W10_HONLY=1`:
-        //   - hot_kv shadow dropped → HOnly (skip K/V readback)
-        //   - hot_kv shadow dropped AND `rs.stored` empty (window=None)
-        //     → None (skip h_in readback too — nothing CPU-side will
-        //     read it)
-        // Backends without optimised paths fall through to Full via
-        // the trait default; correct everywhere, perf-positive on
-        // Metal.
-        let env_on = std::env::var("LARQL_W10_HONLY")
-            .ok()
-            .map(|v| v == "1")
-            .unwrap_or(false);
-        let drop_hot_kv = self
-            .store
-            .as_ref()
-            .map(|s| s.hot_kv.is_none())
-            .unwrap_or(false)
-            && env_on;
-        let drop_stored = self
-            .store
-            .as_ref()
-            .map(|s| s.stored.first().map(|a| a.shape()[0] == 0).unwrap_or(false))
-            .unwrap_or(false)
-            && env_on;
-        let mask = if drop_stored && drop_hot_kv {
-            larql_compute::StateDumpMask::None
-        } else if drop_hot_kv {
-            larql_compute::StateDumpMask::HOnly
-        } else {
-            larql_compute::StateDumpMask::Full
-        };
-        // W10 instrumentation: state_capture = whole backend call
-        // including kernel + state-dump readback. Under HOnly the K/V
-        // readback is skipped; under None all readbacks are skipped.
-        // Comparing this number across mask choices gives the
-        // falsifiable measurement of the kernel-side saving.
-        let t_capture = std::time::Instant::now();
-        let hidden = self.backend.as_ref().coarse_decode_step_with_state_masked(
-            weights,
-            token_id,
-            Some(index),
-            handle,
-            self.abs_position,
-            Some(&mut state),
-            mask,
-        )?;
-        if self.profiling {
-            self.profile.state_capture.record(t_capture);
-        }
-        if !state.is_complete_under(num_layers, mask) {
-            // Backend ran the decode but didn't dump state — engine
-            // can't update its store. Treat as a contract violation;
-            // fall back so the engine's residual store stays
-            // consistent.
-            self.kv_handle = None;
-            return None;
-        }
-        let mut rs = self.store.take()?;
-        // W8.2: append per-layer h_in / K_new / V_new in-place into the
-        // pre-allocated doubling-capacity buffers. `append_row` grows
-        // (Array2::zeros + copy) only on capacity overflow — geometric
-        // growth gives amortised O(1) per token. Replaces the per-step
-        // `Array2::zeros((n+1, dim)) + slice-copy + slice-copy` pattern
-        // that was 58% of decode CPU at 1000 tokens.
-        // W10 Phase A: consume the per-layer handles. CpuStateHandle's
-        // `into_array` moves the inner Array2 out without a copy, so
-        // total memcpy count is unchanged vs pre-W10.
-        // W10 Phase B/C: under HOnly the K/V handle vecs are empty —
-        // we only consume h_in. Under None both are empty — engine
-        // skips append entirely. K/V state lives in Metal's kv cache.
-        let len = rs.hot_len;
-        let h_handles = std::mem::take(&mut state.h_in_per_layer);
-        let k_handles = std::mem::take(&mut state.k_new_per_layer);
-        let v_handles = std::mem::take(&mut state.v_new_per_layer);
-        let did_append = !matches!(mask, larql_compute::StateDumpMask::None);
-        if matches!(mask, larql_compute::StateDumpMask::None) {
-            // Nothing to append on either side; abs_position is the
-            // only piece of canonical state that needs to advance.
-            // `rs.hot_len` deliberately stays at 0 to preserve the
-            // invariant `hot_len == stored[0].shape()[0]`; otherwise
-            // `memory_bytes()` and `window_tokens()` would lie.
-            drop((h_handles, k_handles, v_handles));
-        } else if matches!(mask, larql_compute::StateDumpMask::HOnly) {
-            // Drop K/V handles (vecs are empty under HOnly anyway).
-            drop((k_handles, v_handles));
-            for (layer, h) in h_handles.into_iter().enumerate() {
-                let t_mat = std::time::Instant::now();
-                let h_arr = h.into_array();
-                if self.profiling {
-                    self.profile.state_materialise.record(t_mat);
-                }
-                let t_app = std::time::Instant::now();
-                append_row(&mut rs.stored[layer], &h_arr, len);
-                if self.profiling {
-                    self.profile.state_append.record(t_app);
-                }
-            }
-        } else {
-            for (layer, ((h, k), v)) in h_handles
-                .into_iter()
-                .zip(k_handles)
-                .zip(v_handles)
-                .enumerate()
-            {
-                let t_mat = std::time::Instant::now();
-                let h_arr = h.into_array();
-                let k_arr_opt = if rs.hot_kv.is_some() {
-                    Some((k.into_array(), v.into_array()))
-                } else {
-                    None
-                };
-                if self.profiling {
-                    self.profile.state_materialise.record(t_mat);
-                }
-                let t_app = std::time::Instant::now();
-                append_row(&mut rs.stored[layer], &h_arr, len);
-                if let Some(hot_kv) = rs.hot_kv.as_mut() {
-                    if let Some((k_arr, v_arr)) = k_arr_opt {
-                        append_row(&mut hot_kv[layer].0, &k_arr, len);
-                        append_row(&mut hot_kv[layer].1, &v_arr, len);
-                    }
-                }
-                if self.profiling {
-                    self.profile.state_append.record(t_app);
-                }
-            }
-        }
-        if did_append {
-            rs.hot_len = len + 1;
-        }
-        // Window clip — same snapshot-evicted-into-cold flow as W2.
-        // W8.2: use `rs.hot_len`, not `stored[l].shape()[0]` — with
-        // pre-allocation the latter is the doubling capacity.
-        let pre_clip: Vec<usize> = if rs.hot_kv.is_some() {
-            let window = rs.max_window.unwrap_or(usize::MAX);
-            let evict_count = rs.hot_len.saturating_sub(window);
-            // Same evict count for every layer since they grow in
-            // lockstep. (Pre-W8.2 this read shape[0] per layer; with
-            // hot_len they all share the value.)
-            vec![evict_count; rs.stored.len()]
-        } else {
-            Vec::new()
-        };
-        let evicted_hot_kv = rs
-            .hot_kv
-            .as_ref()
-            .filter(|_| pre_clip.iter().any(|&n| n > 0))
-            .and_then(|h| RsStore::snapshot_evicted_hot_kv(h, &pre_clip));
-        let mut overflow: Vec<Array2<f32>> = Vec::with_capacity(num_layers);
-        for layer in 0..num_layers {
-            rs.clip_layer(layer, &mut overflow);
-        }
-        rs.finalise_hot_len_after_clip();
-        // 2026-05-19 audit fix: geometric-capacity cold append.
-        // Replaces the prior `Array2::zeros((c_old + c_new, ...))`
-        // realloc-and-copy that was O(N) per step / O(N²) total.
-        // See RsStore::append_cold_overflow for the doubling logic.
-        rs.append_cold_overflow(overflow, evicted_hot_kv);
-        self.store = Some(rs);
-        self.abs_position += 1;
-        if self.profiling {
-            self.profile.decode_total.record(t_total);
-        }
-        Some(hidden)
-    }
 }
+
+// W1-GPU dispatch methods (`try_prefill_via_dispatch` /
+// `decode_step_via_dispatch`) live in [`super::dispatch`] as an
+// additional `impl MarkovResidualEngine` block. They mutate the
+// `pub(super)` fields above.
 
 impl KvEngine for MarkovResidualEngine {
     fn name(&self) -> &str {
