+++ crates/larql-kv/src/engines/markov_residual_codec/mod.rs
@@ -14,7 +14,10 @@
 
 pub mod codec;
 pub mod compute;
+pub(crate) mod dispatch;
 pub mod engine;
+pub(crate) mod executor;
+pub(crate) mod helpers;
 pub mod store;
 pub mod walk;
